swagger-backend-docker/requirements.txt
Flask==0.10.1
PyYAML==3.11
python-slugify==1.2.0





